feat: Added new testcase and created a Util for decorators
