+++ packages/concerto-core/lib/decoratormanager.js
@@ -20,6 +20,7 @@ const Factory = require('./factory');
 const ModelUtil = require('./modelutil');
 const { MetaModelNamespace } = require('@accordproject/concerto-metamodel');
 const semver = require('semver');
+const DecoratorUtil = require('./decoratorutil');
 
 // Types needed for TypeScript generation.
 /* eslint-disable no-unused-vars */
@@ -252,14 +253,20 @@ class DecoratorManager {
         newModelManager.fromAst(decoratedAst);
         return newModelManager;
     }
-
+    /**
+     * @typedef ExtractDecoratorsResult
+     * @type {object}
+     * @property {ModelManager} modelManager - A model manager containing models stripped without decorators
+     * @property {decoratorCommandSet} object[] - Stripped out decorators, formed into decorator command sets
+     * @property {vocabularies} object[] - Stripped out vocabularies, formed into vocabulary files
+    */
     /**
      * Extracts all the decorator commands from all the models in modelManager
      * @param {ModelManager} modelManager the input model manager
      * @param {object} options - decorator models options
      * @param {boolean} options.removeDecoratorsFromModel - flag to strip out decorators from models
      * @param {string} options.locale - locale for extracted vocabulary set
-     * @returns {Object} a new model manager with the decorations removed and a list of extracted decorator jsons
+     * @returns {ExtractDecoratorsResult} - a new model manager with the decorations removed and a list of extracted decorator jsons and vocab yamls
      */
     static extractDecorators(modelManager,options) {
         options = {
@@ -271,14 +278,17 @@ class DecoratorManager {
         let extractionDictionary = {};
         const processedModels = decoratedAst.models.map((model)=>{
             if ((model?.decorators.length > 0)){
-                extractionDictionary = this.constructDCSDictionary(extractionDictionary,model.namespace,model.decorators,'','');
+                extractionDictionary = DecoratorUtil.constructDCSDictionary(extractionDictionary,model.namespace,model.decorators, {});
                 if (options.removeDecoratorsFromModel){
                     model.decorators = undefined;
                 }
             }
             const processedDecl = model.declarations.map(decl => {
                 if (decl.decorators) {
-                    extractionDictionary = this.constructDCSDictionary(extractionDictionary,model.namespace,decl.decorators,decl.name,'');
+                    const constructOptions = {
+                        declaration: decl.name,
+                    };
+                    extractionDictionary = DecoratorUtil.constructDCSDictionary(extractionDictionary,model.namespace,decl.decorators,constructOptions);
                 }
                 if (options.removeDecoratorsFromModel){
                     decl.decorators = undefined;
@@ -286,13 +296,21 @@ class DecoratorManager {
                 if (decl.$class.endsWith('.MapDeclaration')) {
                     if (decl.key){
                         if (decl.key.decorators){
-                            extractionDictionary = this.constructDCSDictionary(extractionDictionary,model.namespace,decl.key.decorators,decl.name,'','KEY');
+                            const constructOptions = {
+                                declaration: decl.name,
+                                mapElement: 'KEY'
+                            };
+                            extractionDictionary = DecoratorUtil.constructDCSDictionary(extractionDictionary,model.namespace,decl.key.decorators,constructOptions);
                             decl.key.decorators = undefined;
                         }
                     }
                     if (decl.value){
                         if (decl.value.decorators){
-                            extractionDictionary = this.constructDCSDictionary(extractionDictionary,model.namespace,decl.value.decorators,decl.name,'','VALUE');
+                            const constructOptions = {
+                                declaration: decl.name,
+                                mapElement: 'VALUE'
+                            };
+                            extractionDictionary = DecoratorUtil.constructDCSDictionary(extractionDictionary,model.namespace,decl.value.decorators,constructOptions);
                             decl.value.decorators = undefined;
                         }
                     }
@@ -300,7 +318,11 @@ class DecoratorManager {
                 if (decl.properties) {
                     const processedProperties = decl.properties.map((property) => {
                         if (property.decorators){
-                            extractionDictionary = this.constructDCSDictionary(extractionDictionary,model.namespace, property.decorators,decl.name,property.name);
+                            const constructOptions = {
+                                declaration: decl.name,
+                                property: property.name
+                            };
+                            extractionDictionary = DecoratorUtil.constructDCSDictionary(extractionDictionary ,model.namespace, property.decorators, constructOptions );
                         }
                         if (options.removeDecoratorsFromModel){
                             property.decorators = undefined;
@@ -320,182 +342,13 @@ class DecoratorManager {
         };
         const newModelManager = new ModelManager();
         newModelManager.fromAst(processedAST);
-        const decoratorCommandSet = this.parseNonVocabularyDecorators(extractionDictionary);
-        const vocabularies = this.parseVocabularies(extractionDictionary,options.locale);
+        const decoratorCommandSet = DecoratorUtil.parseNonVocabularyDecorators(extractionDictionary,DCS_VERSION);
+        const vocabularies = DecoratorUtil.parseVocabularies(extractionDictionary,options.locale);
         return {
             modelManager:newModelManager,
             decoratorCommandSet,
             vocabularies
         };
-    }
-    /**
-     * Parses the dict data into an array of decorator jsons
-     * @param {Object} decoratorDict the input dict
-     * @returns {Array<Object>} the parsed decorator command set array
-     * @private
-     */
-    static parseNonVocabularyDecorators(decoratorDict){
-        const data = [];
-        Object.keys(decoratorDict).forEach((namespace)=>{
-            const {name, version} = ModelUtil.parseNamespace(namespace);
-            const nameOfDcs = name;
-            const versionOfDcs = version;
-            const dcsObjects = [];
-            const jsonData = decoratorDict[namespace];
-            const patternToDetermineVocab = /^Term_/i;
-            jsonData.forEach((obj)=>{
-                const decos = JSON.parse(obj.dcs);
-                const target = {
-                    '$class': `org.accordproject.decoratorcommands@${DCS_VERSION}.CommandTarget`,
-                    'namespace':namespace
-                };
-                if (obj.declaration && obj.declaration!==''){
-                    target.declaration = obj.declaration;
-                }
-                if (obj.property && obj.property!==''){
-                    target.property = obj.property;
-                }
-                if (obj.mapElement && obj.mapElement!==''){
-                    target.mapElement = obj.mapElement;
-                }
-                decos.forEach((dcs)=>{
-                    if (dcs.name !== 'Term' && !patternToDetermineVocab.test(dcs.name)){
-                        const decotatorObj = {
-                            '$class': 'concerto.metamodel@1.0.0.Decorator',
-                            'name': dcs.name,
-                        };
-                        if (dcs.arguments){
-                            const args = dcs.arguments.map((arg)=>{
-                                return {
-                                    '$class':arg.$class,
-                                    'value':arg.value
-                                };
-                            });
-                            decotatorObj.arguments = args;
-                        }
-                        let dcsObject = {
-                            '$class': `org.accordproject.decoratorcommands@${DCS_VERSION}.Command`,
-                            'type': 'UPSERT',
-                            'target': target,
-                            'decorator': decotatorObj,
-                        };
-                        dcsObjects.push(dcsObject);
-                    }
-                });
-            });
-            const dcmsForNamespace = {
-                '$class': `org.accordproject.decoratorcommands@${DCS_VERSION}.DecoratorCommandSet`,
-                'name': nameOfDcs,
-                'version': versionOfDcs,
-                'commands': dcsObjects
-            };
-            data.push(dcmsForNamespace);
-        });
-        return data;
-    }
-    /**
-     * Parses the dict data into an array of decorator jsons
-     * @param {Object} decoratorDict the input dict
-     * @param {String} locale locale for target vocab
-     * @returns {Array<Object>} the parsed decorator command set array
-     * @private
-     */
-    static parseVocabularies(decoratorDict,locale){
-        const data = [];
-        const patternToDetermineVocab = /^Term_/i;
-        Object.keys(decoratorDict).forEach((namespace)=>{
-            let strVoc = '';
-            strVoc = strVoc+`locale: ${locale}\n`;
-            strVoc = strVoc+`namespace: ${namespace}\n`;
-            strVoc = strVoc+'declarations:\n';
-            const jsonData = decoratorDict[namespace];
-            const dictVoc = {};
-            jsonData.forEach((obj)=>{
-                if (!dictVoc[obj.declaration]){
-                    dictVoc[obj.declaration] = {
-                        propertyVocabs: {}
-                    };
-                }
-                const decos = JSON.parse(obj.dcs);
-                decos.forEach((dcs)=>{
-                    if (dcs.name === 'Term' || patternToDetermineVocab.test(dcs.name)){
-                        if (obj.property !== ''){
-                            if (!dictVoc[obj.declaration].propertyVocabs[obj.property]){
-                                dictVoc[obj.declaration].propertyVocabs[obj.property] = {};
-                            }
-                            if (dcs.name === 'Term'){
-                                dictVoc[obj.declaration].propertyVocabs[obj.property].term = dcs.arguments[0].value;
-                            }
-                            else {
-                                const extensionKey = dcs.name.split('Term_')[1];
-                                dictVoc[obj.declaration].propertyVocabs[obj.property][extensionKey] = dcs.arguments[0].value;
-                            }
-                        }
-                        else {
-                            if (dcs.name === 'Term'){
-                                dictVoc[obj.declaration].term = dcs.arguments[0].value;
-                            }
-                            else {
-                                const extensionKey = dcs.name.split('Term_')[1];
-                                dictVoc[obj.declaration][extensionKey] = dcs.arguments[0].value;
-                            }
-                        }
-                    }
-                });
-
-            });
-            Object.keys(dictVoc).forEach((decl)=>{
-                if (dictVoc[decl].term){
-                    strVoc += `  - ${decl}: ${dictVoc[decl].term}\n`;
-                    const otherProps = Object.keys(dictVoc[decl]).filter((str)=>str!=='term' && str!=='propertyVocabs');
-                    otherProps.forEach((key)=>{
-                        strVoc += `    ${key}: ${dictVoc[decl][key]}\n`;
-                    });
-                }
-                if (dictVoc[decl].propertyVocabs && Object.keys(dictVoc[decl].propertyVocabs).length>0){
-                    if (!dictVoc[decl].term){
-                        strVoc += `  - ${decl}: ${decl}\n`;
-                    }
-                    strVoc += '    properties:\n';
-                    Object.keys(dictVoc[decl].propertyVocabs).forEach((prop)=>{
-                        strVoc += `      - ${prop}: ${dictVoc[decl].propertyVocabs[prop].term}\n`;
-                        const otherProps = Object.keys(dictVoc[decl].propertyVocabs[prop]).filter((str)=>str!=='term');
-                        otherProps.forEach((key)=>{
-                            strVoc += `        ${key}: ${dictVoc[decl].propertyVocabs[prop][key]}\n`;
-                        });
-                    });
-                }
-            });
-            data.push(strVoc);
-        });
-        return data;
-    }
-    /**
-    * Adds a key-value pair to a dictionary (object) if the key exists,
-    * or creates a new key with the provided value.
-    *
-    * @param {Object} dictionary - The dictionary (object) to which to add the key-value pair.
-    * @param {string} key - The key to add or update.
-    * @param {any} value - The value to add or update.
-    * @param {string} declaration - The target decl.
-    * @param {string} property - The target property.
-    * @param {string} mapElement - The target mapElement.
-    * @returns {Object} - constructed DCS Dict
-    * @private
-    */
-    static constructDCSDictionary(dictionary, key, value, declaration,property,mapElement='') {
-        const val = {
-            declaration,
-            property,
-            mapElement,
-            dcs: JSON.stringify(value),
-        };
-        if (dictionary[key] && Array.isArray( dictionary[key])) {
-            dictionary[key].push(val);
-        } else {
-            dictionary[key] = [val];
-        }
-        return dictionary;
     }
 
     /**
